+++ lti_consumer/outcomes.py
@@ -183,8 +183,9 @@ class OutcomeService:
             log.debug("[LTI]: %s", error_message)
             return response_xml_template.format(**failure_values)
 
-        anon_id = unquote(sourced_id.split(':')[-1])
-        real_user = self.xblock.runtime.service(self, 'user').get_user_by_anonymous_id(anon_id)
+        user_id = unquote(sourced_id.split(':')[-1])
+        real_user = self.xblock.get_lti_1p1_user_from_user_id(user_id)
+
         if not real_user:  # that means we can't save to database, as we do not have real user id.
             failure_values['imsx_messageIdentifier'] = escape(imsx_message_identifier)
             failure_values['imsx_description'] = "User not found."